requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
anthropic==0.89.0
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
openai==2.31.0
google-genai==1.47.0
exa-py
qdrant-client
sentence-transformers
langfuse